+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 =========
-## Unreleased
+## 1.0.0-rc.16 - 2018-03-20
 ### Fixed
 - Incorrect condition when checking instance on ObjectHelper::create
 - Trait conflict with ElementView::runInternal